+++ src/util/pm.js
@@ -14,3 +14,26 @@ export function isCollapsed (pm) {
   if (anchor == null) return false
   return (anchor === head)
 }
+
+export function indexToPos (doc, index) {
+  let pos = 0
+  for (let i = 0; i < index; i++) {
+    const node = doc.maybeChild(i)
+    if (!node) continue
+    pos += node.nodeSize
+  }
+  return pos
+}
+
+export function indexOfId (doc, id) {
+  for (let i = 0, len = doc.childCount; i < len; i++) {
+    const node = doc.child(i)
+    if (!node.attrs || !node.attrs.id) {
+      continue
+    }
+    if (node.attrs.id === id) {
+      return i
+    }
+  }
+  return -1
+}